Fix(folder structure & some styles)
