+++ package.json
@@ -36,15 +36,7 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.2.8",
     "next-auth": "^4.3.1",
-    "next-redux-wrapper": "^7.0.5",
-    "react-redux": "^7.2.8",
     "react-router-dom": "^5.3.0",
-    "redux": "^4.1.2",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-logger": "^3.0.6",
-    "redux-persist": "^6.0.0",
-    "redux-promise-middleware": "^6.1.2",
-    "redux-thunk": "^2.4.1",
     "rxjs": "^7.5.5",
     "xregexp": "^5.1.0",
     "yup": "^0.32.11"
