+++ packages/storefront/components/account/sign-in.tsx
@@ -1,29 +1,38 @@
+import axios from "axios";
 import { ErrorMessage, Field, Form, Formik } from "formik";
 import type { NextComponentType } from "next";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { useDispatch } from "react-redux";
 
-import { loginSchema } from "../schemas/loginSchema";
-import { signin } from "../store/actions/userActions";
+import Breadcrumbs from "../global/breadcrumbs";
+import Button from "../global/button/button";
+
+import { loginSchema } from "../../schemas/loginSchema";
 
 const Signin: NextComponentType = (props) => {
-  const dispatch = useDispatch();
   const router = useRouter();
+  const baseUrl = "http://localhost:3000";
 
-  function handleSignin(data: any) {
-    dispatch(signin(data, router));
+  async function handleSignin(data: any) {
+    try {
+      const response = await axios.post(`${baseUrl}/user-auth/login`, data);
+      localStorage.setItem("userData", JSON.stringify(response.data));
+      window.location.href = "/home";
+    } catch(error: any) {
+      alert(error.response.data.message);
+    }
   }
 
   return (
     <>
+      <Breadcrumbs route={router.asPath} />
       <div className="flex flex-wrap justify-center">
         <div
           className="flex flex-col my-20 py-7 sm:mx-3 md:mx-5"
           style={{ width: " 35rem ", height: "auto", background: "#f3f3f3" }}
         >
-          <h2 className="text-center text-gray-800">Login</h2>
-          <p className="text-center text-gray-500 mx-5">
+          <h2 className="text-3xl mx-3 text-center text-gray-800">Login</h2>
+          <p className="text-center mt-2 mb-6 text-gray-500 mx-5">
             Please login using account detail below.
           </p>
           <div className="sm:m-5 my-3 md:mx-10">
@@ -48,7 +57,7 @@ const Signin: NextComponentType = (props) => {
                     <div className="mb-4">
                       <Field
                         type="phone"
-                        className="w-full p-2 placeholder-gray-800 outline-0"
+                        className="w-full p-2 placeholder-gray-600 outline-0"
                         id="phone"
                         name="phone"
                         placeholder="Phone"
@@ -61,7 +70,7 @@ const Signin: NextComponentType = (props) => {
                     <div className="mb-4">
                       <Field
                         type="password"
-                        className="w-full p-2 placeholder-gray-800 outline-0"
+                        className="w-full p-2 placeholder-gray-600 outline-0"
                         id="password"
                         name="password"
                         placeholder="Password"
@@ -71,14 +80,19 @@ const Signin: NextComponentType = (props) => {
                       </div>
                     </div>
                     <div className="flex flex-wrap sm:justify-end md:justify-between">
-                      <button
-                        type="submit"
-                        className="text-white bg-green-600/100 hover:bg-black rounded text-center py-2 my-2 sm:w-full md:w-1/4"
-                      >
-                        Sign In
-                      </button>
+                    <Button type="submit"
+                      padding="py-2"
+                      margin="my-2"
+                      responsiveProps="sm:w-full md:w-1/4"
+                      bg="bg-green-600/100"
+                      buttonText="Sign In"
+                      onHover="bg-black"
+                      textColor="text-white"
+                      edge="rounded"
+                      />
+                      
                       <div className=" sm:my-0 md:my-3 text-decoration-none">
-                        <Link href="/forgot-password">
+                        <Link href="/account/forgot-password">
                           <a className="text-decoration-none text-gray-600 hover:text-gray-500 font-weight-light">
                             Forgot your password?
                           </a>
@@ -90,7 +104,7 @@ const Signin: NextComponentType = (props) => {
               }}
             </Formik>
             <div className="text-decoration-none mt-3">
-              <Link href="/sign-up">
+              <Link href="/account/sign-up">
                 <a className="text-decoration-none text-gray-600 hover:text-gray-500 font-weight-light">
                   Create account
                 </a>
