+++ packages/storefront/components/global/breadcrumbs/breadcrumbItem.tsx
@@ -4,7 +4,7 @@ const BreadcrumbItem = ({ children, href, ...props }: any) => {
   return (
     <li {...props}>
       <Link href={href} passHref>
-        <a className="text-decoration-none text-black hover:text-green-600/100">{children}</a>
+        <a className="text-decoration-none text-gray-600 hover:text-green-600">{children}</a>
       </Link>
     </li>
   );
