+++ packages/storefront/components/navbar.tsx
@@ -1,10 +1,15 @@
 import type { NextComponentType } from "next";
 import Link from "next/link";
 import { useEffect, useState } from "react";
-import { useSelector } from "react-redux";
+import signout from "./account/sign-out";
 
 const Navbar: NextComponentType = () => {
-  const loggedInUser = useSelector((state: any) => state.userReducer.loggedInUser);
+  const [loggedInUser, setLoggedInUser] = useState('');
+
+  useEffect(() => {
+    setLoggedInUser(localStorage.getItem('userData') || '');
+  }, [])
+
   return (
     <>
       <nav className="navbar navbar-expand-lg navbar-light bg-light fixed-top">
@@ -78,12 +83,14 @@ const Navbar: NextComponentType = () => {
               </button>
               {loggedInUser ? (
                 <>
-                  <Link href="/sign-out">
-                    <button className="btn btn-outline-dark">Sign-Out</button>
+                  <Link href="/account/sign-in">
+                    <button onClick={() => {
+                      signout();
+                    }} className="btn btn-outline-dark">Sign-Out</button>
                   </Link>
                 </>
               ) : (
-                <Link href="/sign-in">
+                <Link href="/account/sign-in">
                   <button className="btn btn-outline-dark">Sign-In</button>
                 </Link>
               )}
