+++ packages/storefront/pages/_app.tsx
@@ -1,25 +1,15 @@
-import "../styles/App.scss";
 import type { AppProps } from "next/app";
 import Layout from "../components/layout";
-import { useEffect } from "react";
-import store from "../store/store";
-import { Provider, useSelector } from "react-redux";
+import "../styles/global.css";
+import "tailwindcss/tailwind.css";
 
 function MyApp({ Component, pageProps }: AppProps) {
 
-  useEffect(() => {
-    typeof document !== undefined
-      ? require("bootstrap/dist/js/bootstrap")
-      : null;
-  }, []);
-
   return (
     <>
-      <Provider store={store}>
         <Layout>
           <Component {...pageProps} />
         </Layout>
-      </Provider>
     </>
   );
 }
