+++ packages/storefront/pages/account/sign-in.tsx
@@ -1,0 +1,11 @@
+import type { GetServerSideProps, NextPage } from "next";
+import Signin from "../../components/account/sign-in";
+
+const SignIn: NextPage = () => {
+  return (
+    <>
+       <Signin/>
+    </>
+  );
+};
+export default SignIn;
