+++ packages/storefront/pages/account/sign-up.tsx
@@ -1,0 +1,11 @@
+import type { GetServerSideProps, NextPage } from "next";
+import Signup from "../../components/account/sign-up";
+
+const SignIn: NextPage = () => {
+  return (
+    <>
+       <Signup/>
+    </>
+  );
+};
+export default SignIn;
