--- packages/storefront/store/reducers/index.ts
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux'
-import userReducer from './userReducer'
-
-export default combineReducers({
-  userReducer
-})
