--- packages/storefront/store/types.ts
@@ -1,7 +1,0 @@
-export default {
-    USER_REGISTER: 'USER_REGISTER',
-
-    USER_LOGIN: 'USER_LOGIN',
-    USER_LOGIN_FULFILLED: 'USER_LOGIN_FULFILLED',
-    USER_LOGOUT: 'USER_LOGOUT',
-}
