--- packages/storefront/styles/App.scss
@@ -1,17 +1,0 @@
-@import "tailwindcss/base";
-
-@import "tailwindcss/components";
-
-@import "tailwindcss/utilities";
-
-@import "~bootstrap/scss/bootstrap";
-#__next {
-  height: 100%;
-  display: flex !important;
-  flex-direction: inherit;
-  font-family: sans-serif;
-  font-weight: 400;
-  font-size: small;
-  letter-spacing: 0.09em;
-}
-
