+++ packages/storefront/tsconfig.json
@@ -15,6 +15,6 @@
     "jsx": "preserve",
     "incremental": true
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/sign-up.jsx", "components/breadcrumbs/breadcrumb.jsx", "components/breadcrumbs/breadcrumbItem.jsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/sign-up.jsx", "components/global/breadcrumbs/breadcrumb.jsx", "components/global/breadcrumbs/breadcrumbItem.jsx"],
   "exclude": ["node_modules"]
 }
